+++ src/parts/array/_findFirstIndex.js
@@ -4,7 +4,7 @@ const _findFirstIndex = (array, func) => {
   for (let i = 0, l = array.length; i < l; i += 1) {
     const resultFunc = func(array[i], i, array);
     if (!isBoolean(resultFunc)) {
-      throw new TypeError(
+      throw new Error(
         `_findFirstIndex args(compareFunc) result is not boolean`,
       );
     }