+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.2.1] - 2024/01/22 Mon
 - update README
 - change Command delimter '|' -> ':'
+- update fix misstake Error
 
 ## [1.2.0] - 2024/01/22 Mon
 - refactoring